lavc: move default get_buffer2() to its own file
It is also used by some encoders, so decode.c is not the right place for
it.
